+++ R/bootstrap.R
@@ -101,6 +101,8 @@
 #'                                are set to their values at time 0. The default is \code{FALSE}.
 #' @param below_zero_indicator    Logical scalar indicating whether the observed data set contains rows for time \eqn{t < 0}.
 #' @param min_time                Numeric scalar specifying lowest value of time \eqn{t} in the observed data set.
+#' @param show_progress           Logical scalar indicating whether to print a progress bar for the number of bootstrap samples completed in the R console. This argument is only applicable when \code{parallel} is set to \code{FALSE} and bootstrap samples are used (i.e., \code{nsamples} is set to a value greater than 0). The default is \code{FALSE}.
+#' @param pb                      Progress bar R6 object. See \code{\link[progress]{progress_bar}} for further details.
 #' @return                        A list with the following components:
 #' \item{Result}{Matrix containing risks over time under the natural course and under each user-specific intervention.}
 #' \item{ResultRatio}{Matrix containing risk ratios over time under the natural course and under each user-specific intervention.}
@@ -118,7 +120,7 @@ bootstrap_helper <- function(r, time_points, obs_data, bootseeds, outcome_type,
                              time_name, outcome_name, compevent_name,
                              ranges, yrange, compevent_range, parallel, ncores, max_visits,
                              hazardratio, intcomp, boot_diag, nsimul, baselags,
-                             below_zero_indicator, min_time){
+                             below_zero_indicator, min_time, show_progress, pb){
 
   set.seed(bootseeds[r])
 
@@ -188,7 +190,7 @@ bootstrap_helper <- function(r, time_points, obs_data, bootseeds, outcome_type,
              subseed = bootseeds[r], time_points = time_points,
              obs_data = resample_data, parallel = FALSE, max_visits = max_visits,
              baselags = baselags, below_zero_indicator = below_zero_indicator,
-             min_time = min_time)
+             min_time = min_time, show_progress = show_progress, pb = pb)
   })
 
   nat_pool <- pools[[1]] # Simulated data under natural course