+++ R/gformula.R
@@ -101,6 +101,7 @@
 #' @param ci_method               Character string specifying the method for calculating the bootstrap 95\% confidence intervals, if applicable. The options are \code{"percentile"} and \code{"normal"}.
 #' @param threads                 Integer specifying the number of threads to be used in \code{data.table}. See \code{\link[data.table]{setDTthreads}} for further details.
 #' @param boot_diag               Logical scalar indicating whether to return the coefficients of the fitted models and their standard errors in the bootstrap samples. The default is \code{FALSE}.
+#' @param show_progress           Logical scalar indicating whether to print a progress bar for the number of bootstrap samples completed in the R console. This argument is only applicable when \code{parallel} is set to \code{FALSE} and bootstrap samples are used (i.e., \code{nsamples} is set to a value greater than 0). The default is \code{FALSE}.
 #' @param ...                     Other arguments, which are passed to the functions in \code{covpredict_custom}.
 #' @return                        An object of class \code{gformula_survival}. The object is a list with the following components:
 #' \item{result}{Results table. For survival outcomes, this contains the estimated risk, risk difference, and risk ratio for all interventions (inculding the natural course) at each time point. For continuous end-of-follow-up outcomes, this contains estimated mean outcome, mean difference, and mean ratio for all interventions (inculding natural course) at the last time point. For binary end-of-follow-up outcomes, this contains the estimated outcome probability, probability difference, and probability ratio for all interventions (inculding natural course) at the last time point. If bootstrapping was used, the results table includes the bootstrap risk / mean / probability difference, ratio, standard error, and 95\% confidence interval.}
@@ -244,6 +245,7 @@
 #' ## Estimating the effect of threshold interventions on the mean of a binary
 #' ## end of follow-up outcome
 #' \donttest{
+#' outcome_type <- 'binary_eof'
 #' id <- 'id_num'
 #' time_name <- 'time'
 #' covnames <- c('cov1', 'cov2', 'treat')
@@ -251,10 +253,10 @@
 #' histories <- c(lagged, cumavg)
 #' histvars <- list(c('treat', 'cov1', 'cov2'), c('cov1', 'cov2'))
 #' covtypes <- c('binary', 'zero-inflated normal', 'normal')
-#' covparams <- list(covmodels = c(cov1 ~ lag1_treat + lag1_cov1 + lag1_cov2 + cov3 +
-#'                                   time,
-#'                                 cov2 ~ lag1_treat + cov1 + lag1_cov1 + lag1_cov2 +
+#' covparams <- list(covmodels = c(cov1 ~ lag1_treat + lag1_cov1 + lag1_cov2 +
 #'                                   cov3 + time,
+#'                                 cov2 ~ lag1_treat + cov1 + lag1_cov1 +
+#'                                   lag1_cov2 + cov3 + time,
 #'                                 treat ~ lag1_treat + cumavg_cov1 +
 #'                                   cumavg_cov2 + cov3 + time))
 #' ymodel <- outcome ~  treat + cov1 + cov2 + lag1_cov1 + lag1_cov2 + cov3
@@ -265,20 +267,16 @@
 #' nsimul <- 10000
 #' ncores <- 2
 #'
-#' gform_bin_eof <- gformula_binary_eof(obs_data = binary_eofdata, id = id,
-#'                                      time_name = time_name,
-#'                                      covnames = covnames,
-#'                                      outcome_name = outcome_name,
-#'                                      covtypes = covtypes,
-#'                                      covparams = covparams,
-#'                                      ymodel = ymodel,
-#'                                      intvars = intvars,
-#'                                      interventions = interventions,
-#'                                      int_descript = int_descript,
-#'                                      histories = histories, histvars = histvars,
-#'                                      basecovs = c("cov3"), seed = 1234,
-#'                                      parallel = TRUE, nsamples = 5,
-#'                                      nsimul = nsimul, ncores = ncores)
+#' gform_bin_eof <- gformula(obs_data = binary_eofdata,
+#'                           outcome_type = outcome_type, id = id,
+#'                           time_name = time_name, covnames = covnames,
+#'                           outcome_name = outcome_name, covtypes = covtypes,
+#'                           covparams = covparams, ymodel = ymodel,
+#'                           intvars = intvars, interventions = interventions,
+#'                           int_descript = int_descript, histories = histories,
+#'                           histvars = histvars, basecovs = c("cov3"),
+#'                           seed = 1234, parallel = TRUE, nsamples = 5,
+#'                           nsimul = nsimul, ncores = ncores)
 #' gform_bin_eof
 #' }
 #'
@@ -299,7 +297,7 @@ gformula <- function(obs_data, id, time_points = NULL,
                      nsimul = NA, sim_data_b = FALSE, seed,
                      nsamples = 0, parallel = FALSE, ncores = NA,
                      ci_method = 'percentile', threads,
-                     boot_diag = FALSE, ...){
+                     boot_diag = FALSE, show_progress = TRUE, ...){
   if (! outcome_type %in% c('survival', 'continuous_eof', 'binary_eof')){
     stop("outcome_type must be 'survival', 'continuous_eof', or 'binary_eof', but outcome_type was set to", outcome_type)
   }
@@ -324,7 +322,7 @@ gformula <- function(obs_data, id, time_points = NULL,
                       nsimul = nsimul, sim_data_b = sim_data_b, seed = seed,
                       nsamples = nsamples, parallel = parallel, ncores = ncores,
                       ci_method = ci_method, threads = threads,
-                      boot_diag = boot_diag, ...)
+                      boot_diag = boot_diag, show_progress = show_progress, ...)
   } else if (outcome_type == 'continuous_eof'){
     gformula_continuous_eof(obs_data = obs_data, id = id,
                             time_name = time_name, covnames = covnames,
@@ -345,7 +343,8 @@ gformula <- function(obs_data, id, time_points = NULL,
                             seed = seed, nsamples = nsamples,
                             parallel = parallel, ncores = ncores,
                             ci_method = ci_method, threads = threads,
-                            boot_diag = boot_diag, ...)
+                            boot_diag = boot_diag,
+                            show_progress = show_progress, ...)
   } else if (outcome_type == 'binary_eof'){
     gformula_binary_eof(obs_data = obs_data, id = id,
                         time_name = time_name, covnames = covnames,
@@ -364,7 +363,8 @@ gformula <- function(obs_data, id, time_points = NULL,
                         nsamples = nsamples, parallel = parallel,
                         ncores = ncores,
                         ci_method = ci_method, threads = threads,
-                        boot_diag = boot_diag, ...)
+                        boot_diag = boot_diag, show_progress = show_progress,
+                        ...)
   }
 }
 
@@ -470,6 +470,7 @@ gformula <- function(obs_data, id, time_points = NULL,
 #' @param ci_method               Character string specifying the method for calculating the bootstrap 95\% confidence intervals, if applicable. The options are \code{"percentile"} and \code{"normal"}.
 #' @param threads                 Integer specifying the number of threads to be used in \code{data.table}. See \code{\link[data.table]{setDTthreads}} for further details.
 #' @param boot_diag               Logical scalar indicating whether to return the coefficients of the fitted models and their standard errors in the bootstrap samples. The default is \code{FALSE}.
+#' @param show_progress           Logical scalar indicating whether to print a progress bar for the number of bootstrap samples completed in the R console. This argument is only applicable when \code{parallel} is set to \code{FALSE} and bootstrap samples are used (i.e., \code{nsamples} is set to a value greater than 0). The default is \code{FALSE}.
 #' @param ...                     Other arguments, which are passed to the functions in \code{covpredict_custom}.
 #' @return                        An object of class \code{gformula_survival}. The object is a list with the following components:
 #' \item{result}{Results table containing the estimated risk and risk ratio for all interventions (inculding the natural course) at each time point. If bootstrapping was used, the results table includes the bootstrap mean risk ratio, standard error, and 95\% confidence interval.}
@@ -587,7 +588,7 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
                               nsimul = NA, sim_data_b = FALSE, seed,
                               nsamples = 0, parallel = FALSE, ncores = NA,
                               ci_method = 'percentile', threads,
-                              boot_diag = FALSE, ...){
+                              boot_diag = FALSE, show_progress = TRUE, ...){
 
   lag_indicator <- lagavg_indicator <- cumavg_indicator <- c()
   lag_indicator <- update_lag_indicator(covparams$covmodels, lag_indicator)
@@ -798,7 +799,7 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
                                  subseed = subseed, time_points = time_points,
                                  obs_data = obs_data, parallel = parallel, max_visits = max_visits,
                                  baselags = baselags, below_zero_indicator = below_zero_indicator,
-                                 min_time = min_time, ...)
+                                 min_time = min_time, show_progress = FALSE, ...)
     parallel::stopCluster(cl)
   } else {
     pools <- lapply(seq_along(comb_interventions), FUN = function(i){
@@ -818,7 +819,7 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
                subseed = subseed, time_points = time_points,
                obs_data = obs_data, parallel = parallel, max_visits = max_visits,
                baselags = baselags, below_zero_indicator = below_zero_indicator,
-               min_time = min_time, ...)
+               min_time = min_time, show_progress = FALSE, ...)
     })
   }
 
@@ -903,9 +904,15 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
                                       compevent_name = compevent_name, parallel = parallel, ncores = ncores,
                                       max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                                       boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
-                                      below_zero_indicator = below_zero_indicator, min_time = min_time, ...)
+                                      below_zero_indicator = below_zero_indicator, min_time = min_time,
+                                      show_progress = FALSE, ...)
       parallel::stopCluster(cl)
     } else {
+      if (show_progress){
+        pb <- progress::progress_bar$new(total = nsamples * length(comb_interventions),
+                                         clear = FALSE,
+                                         format = 'Bootstrap progress [:bar] :percent, Elapsed time :elapsed, Est. time remaining :eta')
+      }
       final_bs <- lapply(1:nsamples, FUN = bootstrap_helper, time_points = time_points,
                          obs_data = obs_data_noresample, bootseeds = bootseeds,
                          intvars = comb_intvars, interventions = comb_interventions, int_times = comb_int_times, ref_int = ref_int,
@@ -923,7 +930,8 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
                          compevent_name = compevent_name, parallel = parallel, ncores = ncores,
                          max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                          boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
-                         below_zero_indicator = below_zero_indicator, min_time = min_time, ...)
+                         below_zero_indicator = below_zero_indicator, min_time = min_time,
+                         show_progress = show_progress, pb = pb, ...)
     }
 
     comb_result <- rbindlist(lapply(final_bs, FUN = function(m){
@@ -1228,6 +1236,7 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
 #' @param ci_method               Character string specifying the method for calculating the bootstrap 95\% confidence intervals, if applicable. The options are \code{"percentile"} and \code{"normal"}.
 #' @param threads                 Integer specifying the number of threads to be used in \code{data.table}. See \code{\link[data.table]{setDTthreads}} for further details.
 #' @param boot_diag               Logical scalar indicating whether to return the coefficients of the fitted models and their standard errors in the bootstrap samples. The default is \code{FALSE}.
+#' @param show_progress           Logical scalar indicating whether to print a progress bar for the number of bootstrap samples completed in the R console. This argument is only applicable when \code{parallel} is set to \code{FALSE} and bootstrap samples are used (i.e., \code{nsamples} is set to a value greater than 0). The default is \code{FALSE}.
 #' @param ...                     Other arguments, which are passed to the functions in \code{covpredict_custom}.
 #'
 #' @return                        An object of class \code{gformula_continuous_eof}. The object is a list with the following components:
@@ -1300,7 +1309,8 @@ gformula_continuous_eof <- function(obs_data, id,
                                     sim_data_b = FALSE,  seed, nsamples = 0,
                                     parallel = FALSE, ncores = NA,
                                     ci_method = 'percentile', threads,
-                                    boot_diag = FALSE, ...){
+                                    boot_diag = FALSE, show_progress = TRUE,
+                                    ...){
 
   lag_indicator <- lagavg_indicator <- cumavg_indicator <- c()
   lag_indicator <- update_lag_indicator(covparams$covmodels, lag_indicator)
@@ -1502,7 +1512,7 @@ gformula_continuous_eof <- function(obs_data, id,
                                  subseed = subseed, time_points = time_points,
                                  obs_data = obs_data, parallel = parallel,
                                  baselags = baselags, below_zero_indicator = below_zero_indicator,
-                                 min_time = min_time, ...)
+                                 min_time = min_time, show_progress = FALSE, ...)
     parallel::stopCluster(cl)
   } else {
     pools <- lapply(seq_along(comb_interventions), FUN = function(i){
@@ -1521,7 +1531,7 @@ gformula_continuous_eof <- function(obs_data, id,
                subseed = subseed, time_points = time_points,
                obs_data = obs_data, parallel = parallel,
                baselags = baselags, below_zero_indicator = below_zero_indicator,
-               min_time = min_time, ...)
+               min_time = min_time, show_progress = FALSE, ...)
     })
   }
 
@@ -1571,10 +1581,16 @@ gformula_continuous_eof <- function(obs_data, id,
                                       compevent_name = compevent_name, parallel = parallel, ncores = ncores,
                                       max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                                       boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
-                                      below_zero_indicator = below_zero_indicator, min_time = min_time, ...)
+                                      below_zero_indicator = below_zero_indicator, min_time = min_time,
+                                      show_progress = FALSE, ...)
       parallel::stopCluster(cl)
 
     } else {
+      if (show_progress){
+        pb <- progress::progress_bar$new(total = nsamples * length(comb_interventions),
+                                         clear = FALSE,
+                                         format = 'Bootstrap progress [:bar] :percent, Elapsed time :elapsed, Est. time remaining :eta')
+      }
       final_bs <- lapply(1:nsamples, FUN = bootstrap_helper, time_points = time_points,
                          obs_data = obs_data_noresample, bootseeds = bootseeds,
                          intvars = comb_intvars, interventions = comb_interventions, int_times = comb_int_times, ref_int = ref_int,
@@ -1592,7 +1608,8 @@ gformula_continuous_eof <- function(obs_data, id,
                          compevent_name = compevent_name, parallel = parallel, ncores = ncores,
                          max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                          boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
-                         below_zero_indicator = below_zero_indicator, min_time = min_time, ...)
+                         below_zero_indicator = below_zero_indicator, min_time = min_time,
+                         show_progress = show_progress, pb = pb, ...)
     }
     comb_result <- rbindlist(lapply(final_bs, FUN = function(m){
       as.data.table(t(m$Result))
@@ -1863,6 +1880,7 @@ gformula_continuous_eof <- function(obs_data, id,
 #' @param ci_method               Character string specifying the method for calculating the bootstrap 95\% confidence intervals, if applicable. The options are \code{"percentile"} and \code{"normal"}.
 #' @param threads                 Integer specifying the number of threads to be used in \code{data.table}. See \code{\link[data.table]{setDTthreads}} for further details.
 #' @param boot_diag               Logical scalar indicating whether to return the coefficients of the fitted models and their standard errors in the bootstrap samples. The default is \code{FALSE}.
+#' @param show_progress           Logical scalar indicating whether to print a progress bar for the number of bootstrap samples completed in the R console. This argument is only applicable when \code{parallel} is set to \code{FALSE} and bootstrap samples are used (i.e., \code{nsamples} is set to a value greater than 0). The default is \code{FALSE}.
 #' @param ...                     Other arguments, which are passed to the functions in \code{covpredict_custom}.
 #'
 #' @return An object of class \code{gformula_binary_eof}. The object is a list with the following components:
@@ -1936,7 +1954,7 @@ gformula_binary_eof <- function(obs_data, id,
                                 nsimul = NA, sim_data_b = FALSE, seed,
                                 nsamples = 0, parallel = FALSE, ncores = NA,
                                 ci_method = 'percentile', threads,
-                                boot_diag = FALSE, ...){
+                                boot_diag = FALSE, show_progress = TRUE, ...){
 
   lag_indicator <- lagavg_indicator <- cumavg_indicator <- c()
   lag_indicator <- update_lag_indicator(covparams$covmodels, lag_indicator)
@@ -2133,7 +2151,7 @@ gformula_binary_eof <- function(obs_data, id,
                                  subseed = subseed, time_points = time_points,
                                  obs_data = obs_data, parallel = parallel,
                                  baselags = baselags, below_zero_indicator = below_zero_indicator,
-                                 min_time = min_time, ...)
+                                 min_time = min_time, show_progress = FALSE, ...)
     parallel::stopCluster(cl)
 
   } else {
@@ -2153,7 +2171,7 @@ gformula_binary_eof <- function(obs_data, id,
                subseed = subseed, time_points = time_points,
                obs_data = obs_data, parallel = parallel,
                baselags = baselags, below_zero_indicator = below_zero_indicator,
-               min_time = min_time, ...)
+               min_time = min_time, show_progress = FALSE, ...)
     })
   }
 
@@ -2203,10 +2221,16 @@ gformula_binary_eof <- function(obs_data, id,
                                       compevent_name = compevent_name, parallel = parallel, ncores = ncores,
                                       max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                                       boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
-                                      below_zero_indicator = below_zero_indicator, min_time = min_time, ...)
+                                      below_zero_indicator = below_zero_indicator, min_time = min_time,
+                                      show_progress = FALSE, ...)
       parallel::stopCluster(cl)
 
     } else {
+      if (show_progress){
+        pb <- progress::progress_bar$new(total = nsamples * length(comb_interventions),
+                                         clear = FALSE,
+                                         format = 'Bootstrap progress [:bar] :percent, Elapsed time :elapsed, Est. time remaining :eta')
+      }
       final_bs <- lapply(1:nsamples, FUN = bootstrap_helper, time_points = time_points,
                          obs_data = obs_data_noresample, bootseeds = bootseeds,
                          intvars = comb_intvars, interventions = comb_interventions, int_times = comb_int_times, ref_int = ref_int,
@@ -2224,7 +2248,8 @@ gformula_binary_eof <- function(obs_data, id,
                          compevent_name = compevent_name, parallel = parallel, ncores = ncores,
                          max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                          boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
-                         below_zero_indicator = below_zero_indicator, min_time = min_time, ...)
+                         below_zero_indicator = below_zero_indicator, min_time = min_time,
+                         show_progress = show_progress, pb = pb, ...)
     }
     comb_result <- rbindlist(lapply(final_bs, FUN = function(m){
       as.data.table(t(m$Result))