+++ R/simulate.R
@@ -129,6 +129,8 @@ predict_trunc_normal <- function(x, mean, est_sd, a, b){
 #'                                are set to their values at time 0. The default is \code{FALSE}.
 #' @param below_zero_indicator    Logical scalar indicating whether the observed data set contains rows for time \eqn{t < 0}.
 #' @param min_time                Numeric scalar specifying lowest value of time \eqn{t} in the observed data set.
+#' @param show_progress           Logical scalar indicating whether to print a progress bar for the number of bootstrap samples completed in the R console. This argument is only applicable when \code{parallel} is set to \code{FALSE} and bootstrap samples are used (i.e., \code{nsamples} is set to a value greater than 0). The default is \code{FALSE}.
+#' @param pb                      Progress bar R6 object. See \code{\link[progress]{progress_bar}} for further details.
 #' @param ...                     Other arguments, which are passed to the functions in \code{covpredict_custom}.
 #' @return                        A data table containing simulated data under the specified intervention.
 #' @keywords internal
@@ -141,7 +143,7 @@ simulate <- function(o, fitcov, fitY, fitD,
                      outcome_type, subseed, obs_data, time_points, parallel,
                      covnames, covtypes, covparams, covpredict_custom,
                      basecovs, max_visits, baselags, below_zero_indicator,
-                     min_time, ...){
+                     min_time, show_progress, pb, ...){
   set.seed(subseed)
 
   # Mechanism of passing intervention variable and intervention is different for parallel
@@ -449,7 +451,7 @@ simulate <- function(o, fitcov, fitY, fitD,
         }
         if (covtypes[i] == 'normal' || covtypes[i] == 'bounded normal' ||
            covtypes[i] == 'truncated normal'){
-  
+
              if (length(newdf[newdf[[covnames[i]]] < ranges[[i]][1]][[covnames[i]]]) != 0){
                 newdf[newdf[[covnames[i]]] < ranges[[i]][1], (covnames[i]) := cast(ranges[[i]][1])]
              }
@@ -603,5 +605,8 @@ simulate <- function(o, fitcov, fitY, fitD,
     pool[, 'survival' := stats::ave(pool$prodp0, by = pool$id, FUN = cumprod)]
   }
   pool2 <- copy(pool)
+  if (show_progress){
+    pb$tick()
+  }
   return (pool2)
 }